LayoutTests/imported/w3c/web-platform-tests/trusted-types/no-require-trusted-types-for-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes

PASS script.src without trusted types
PASS div.innerHTML without trusted types
PASS script.text without trusted types
